+++ c/dynmat.h
@@ -37,23 +37,23 @@
 
 #define PHPYCONST
 
-int dym_get_dynamical_matrix_at_q(double *dynamical_matrix,
-                                  const int num_patom,
-                                  const int num_satom,
-                                  const double *fc,
-                                  const double q[3],
-                                  PHPYCONST double (*svecs)[27][3],
-                                  const int *multi,
-                                  const double *mass,
-                                  const int *s2p_map,
-                                  const int *p2s_map,
-                                  PHPYCONST double (*charge_sum)[3][3],
-                                  const int with_openmp);
+long dym_get_dynamical_matrix_at_q(double *dynamical_matrix,
+                                   const long num_patom,
+                                   const long num_satom,
+                                   const double *fc,
+                                   const double q[3],
+                                   PHPYCONST double (*svecs)[27][3],
+                                   const long *multi,
+                                   const double *mass,
+                                   const long *s2p_map,
+                                   const long *p2s_map,
+                                   PHPYCONST double (*charge_sum)[3][3],
+                                   const long with_openmp);
 void dym_get_recip_dipole_dipole(double *dd, /* [natom, 3, natom, 3, (real,imag)] */
                                  const double *dd_q0, /* [natom, 3, 3, (real,imag)] */
                                  PHPYCONST double (*G_list)[3], /* [num_G, 3] */
-                                 const int num_G,
-                                 const int num_patom,
+                                 const long num_G,
+                                 const long num_patom,
                                  const double q_cart[3],
                                  const double *q_direction_cart, /* must be pointer */
                                  PHPYCONST double (*born)[3][3],
@@ -64,15 +64,15 @@ void dym_get_recip_dipole_dipole(double *dd, /* [natom, 3, natom, 3, (real,imag)
                                  const double tolerance);
 void dym_get_recip_dipole_dipole_q0(double *dd_q0, /* [natom, 3, 3, (real,imag)] */
                                     PHPYCONST double (*G_list)[3], /* [num_G, 3] */
-                                    const int num_G,
-                                    const int num_patom,
+                                    const long num_G,
+                                    const long num_patom,
                                     PHPYCONST double (*born)[3][3],
                                     PHPYCONST double dielectric[3][3],
                                     PHPYCONST double (*pos)[3], /* [natom, 3] */
                                     const double lambda,
                                     const double tolerance);
 void dym_get_charge_sum(double (*charge_sum)[3][3],
-                        const int num_patom,
+                        const long num_patom,
                         const double factor,
                         const double q_cart[3],
                         PHPYCONST double (*born)[3][3]);
@@ -85,11 +85,11 @@ void dym_transform_dynmat_to_fc(double *fc,
                                 const double *dm,
                                 PHPYCONST double (*comm_points)[3],
                                 PHPYCONST double (*shortest_vectors)[27][3],
-                                const int *multiplicities,
+                                const long *multiplicities,
                                 const double *masses,
-                                const int *s2pp_map,
-                                const int *fc_index_map,
-                                const int num_patom,
-                                const int num_satom);
+                                const long *s2pp_map,
+                                const long *fc_index_map,
+                                const long num_patom,
+                                const long num_satom);
 
 #endif
